fix(trusty-review): call resolve_index() in cmd_run/cmd_serve/cmd_compare so auto-derive takes effect (closes #670)
#666 (b1c4710) added ReviewConfig::resolve_index() and the
config/index_resolver module but none of the three production entry
points called it.  When TRUSTY_SEARCH_INDEX is unset the search_index
field stayed at its "main" default for the entire lifetime of the
process, causing 404s on every project that registered its index under
a different id.

Wiring points:
- commands/run.rs  cmd_run: build HttpSearchClient → resolve_index →
  build_deps_async (the provider.clone() avoids a borrow-across-mut).
- commands/serve.rs  build_app_state: construct search client →
  resolve_index → AppState::with_verifier_and_dedup; config param
  changed to mut.
- commands/compare.rs  cmd_compare: resolve once before the per-model
  loop so all runs share the resolved index; config param changed to
  mut.

In all three paths resolve_index is a no-op when search_index_explicit
is true (TRUSTY_SEARCH_INDEX was set by the operator).

Test additions:
- config/config_resolve_index_tests.rs (new file, split from
  config_tests.rs to stay under the 500-line cap):
  - wiring_cmd_run_resolve_index_updates_before_pipeline
  - wiring_build_app_state_explicit_index_unchanged
  - wiring_cmd_compare_resolve_index_applies_to_all_model_runs
  (plus the four existing resolve_index tests migrated here)

All 601 tests pass; clippy clean; line-cap 0 violations.
